cmd/vfkit: add --pidfile flag to write vfkit process PID to file
Introduce a new --pidfile flag to optionally specify a file where the
running vfkit process writes its PID. This allows easier process
management and monitoring.

+++ pkg/process/process.go
@@ -55,7 +55,7 @@ func (p *Process) FindProcess() (*process.Process, error) {
 		return nil, err
 	}
 
-	proc, err := process.NewProcess(int32(pid))
+	proc, err := process.NewProcess(pid)
 	if err != nil && err.Error() == "process does not exist" {
 		return nil, os.ErrNotExist
 	}
